+++ packages/artifact-tool/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @opengeni/artifact-tool
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [1c78ed0]
+- Updated dependencies [f4afa19]
+- Updated dependencies [8583779]
+- Updated dependencies [79ee99b]
+- Updated dependencies [2cb04e0]
+- Updated dependencies [6d22ab5]
+  - @opengeni/contracts@2.0.0
+
 ## 0.3.0
 
 ### Minor Changes